backend-folder copy/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==4.2.20
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
